+++ postgres/tests/postgres.rs
@@ -1,52 +1,15 @@
+use std::collections::HashMap;
 use std::env;
-use std::path::Path;
+use std::time::Duration;
 
 use futures::future::join_all;
 use tokio_postgres::types::Type;
 
-use deadpool_postgres::{Manager, Pool};
-
-fn pg_config_from_env() -> tokio_postgres::config::Config {
-    let mut config = tokio_postgres::Config::new();
-    if let Ok(host) = env::var("PG_HOST") {
-        config.host(host.as_str());
-    } else if Path::new("/run/postgresql").exists() {
-        config.host("/run/postgresql");
-    } else {
-        config.host("/tmp");
-    }
-    if let Ok(port) = env::var("PG_PORT") {
-        match u16::from_str_radix(port.as_str(), 10) {
-            Ok(port) => {
-                config.port(port);
-            }
-            Err(_) => {
-                panic!(format!("Invalid port: {}", port));
-            }
-        }
-    }
-    if let Ok(user) = env::var("PG_USER") {
-        config.user(user.as_str());
-    } else if let Ok(user) = env::var("USER") {
-        config.user(user.as_str());
-    } else {
-        panic!("PG_USER missing in environment; fallback to USER failed as well.");
-    }
-    if let Ok(password) = env::var("PG_PASSWORD") {
-        config.password(password.as_str());
-    }
-    if let Ok(dbname) = env::var("PG_DBNAME") {
-        config.dbname(dbname.as_str());
-    } else {
-        config.dbname("deadpool");
-    }
-    config
-}
+use deadpool_postgres::{Config, Pool};
 
 fn create_pool() -> Pool {
-    let cfg = pg_config_from_env();
-    let mgr = Manager::new(cfg, tokio_postgres::NoTls);
-    Pool::new(mgr, 16)
+    let cfg = Config::from_env("PG").unwrap();
+    cfg.create_pool(tokio_postgres::NoTls)
 }
 
 #[tokio::main]
@@ -140,4 +103,60 @@ async fn test_transaction_pipeline() {
     for i in 0..100 {
         assert_eq!(results[i], (i as i32) + 1);
     }
+}
+
+struct Env {
+    backup: HashMap<String, Option<String>>,
+}
+
+impl Env {
+    pub fn new() -> Self {
+        Self {
+            backup: HashMap::new()
+        }
+    }
+    pub fn set(&mut self, name: &str, value: &str) {
+        self.backup.insert(name.to_string(), env::var(name).ok());
+        env::set_var(name, value);
+    }
+}
+
+impl Drop for Env {
+    fn drop(&mut self) {
+        for (name, value) in self.backup.iter() {
+            println!("setting {} = {:?}", name, value);
+            match value {
+                Some(value) => env::set_var(name.as_str(), value),
+                None => env::remove_var(name.as_str()),
+            }
+        }
+    }
+}
+
+#[test]
+#[cfg(feature = "config")]
+fn test_config_from_env() {
+    // This test must not use "PG" as prefix as this can cause the other
+    // tests which also use the "PG" prefix to fail.
+    let mut env = Env::new();
+    env.set("PG_ENV_TEST_HOST", "pg.example.com");
+    env.set("PG_ENV_TEST_USER", "john_doe");
+    env.set("PG_ENV_TEST_PASSWORD", "topsecret");
+    env.set("PG_ENV_TEST_DBNAME", "example");
+    env.set("PG_ENV_TEST_POOL.MAX_SIZE", "42");
+    env.set("PG_ENV_TEST_POOL.TIMEOUTS.WAIT.SECS", "1");
+    env.set("PG_ENV_TEST_POOL.TIMEOUTS.WAIT.NANOS", "0");
+    env.set("PG_ENV_TEST_POOL.TIMEOUTS.CREATE.SECS", "2");
+    env.set("PG_ENV_TEST_POOL.TIMEOUTS.CREATE.NANOS", "0");
+    env.set("PG_ENV_TEST_POOL.TIMEOUTS.RECYCLE.SECS", "3");
+    env.set("PG_ENV_TEST_POOL.TIMEOUTS.RECYCLE.NANOS", "0");
+    let cfg = Config::from_env("PG_ENV_TEST").unwrap();
+    // FIXME add tests for db connection parameters
+    //let pg_cfg = cfg.get_pg_config();
+    //assert_eq!(pg_cfg.host, vec![tokio_postgres::config::Host::Tcp("pg.example.com")]);
+    let pool_cfg = cfg.get_pool_config();
+    assert_eq!(pool_cfg.max_size, 42);
+    assert_eq!(pool_cfg.timeouts.wait, Some(Duration::from_secs(1)));
+    assert_eq!(pool_cfg.timeouts.create, Some(Duration::from_secs(2)));
+    assert_eq!(pool_cfg.timeouts.recycle, Some(Duration::from_secs(3)));
 }