+++ redis/CHANGELOG.md
@@ -1,10 +1,11 @@
 # Change Log
 
-## v0.5.0 (unreleased)
+## v0.5.0
 
 * Update `redis` dependency to version 0.14.0
 * Rename `query` to `query_async` to match API of `redis` crate
 * Rename `execute` to `execute_async` to match API of `redis` crate
+* Add support for `config` crate
 
 ## v0.4.1